cast is nicer

+++ lsystems_v7.c
@@ -521,7 +521,7 @@ void _lsys_execute_symbol(const lsys_t* lsys,
 
             int delta = (symbol == '+') ? 1 : -1;
 
-            int t = positive_mod(state->angle + delta,
+            int t = positive_mod((int)state->angle + delta,
                                  lsys->rotation_cycle_length);
 
             state->angle = t;
